requirements.txt
fastapi>=0.124.1,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
websockets>=15.0.1,<16.0.0
google-adk==1.27.0
google-genai
google-cloud-firestore==2.19.0
python-dotenv==1.0.1
pydantic>=2.12.0,<3.0.0
httpx==0.28.1
Pillow==11.1.0
PyMuPDF==1.26.3
python-multipart==0.0.20
